+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+## 0.2.1 (2018-02-24)
+
 - [PR [#10](https://github.com/palkan/isolator/pull/10)] Add `sucker_punch` adapter. ([@alexshgov][])
 - [PR [#9](https://github.com/palkan/isolator/pull/9)] Add `resque scheduler` background job adapter. ([@dsalahutdinov][])